Move api helpers to new classes
